Fixed Orientation code duplication and renamed url_contant.dart
