+++ lib/view/settings_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+//import 'package:flutter/services.dart';
 import 'package:magic_epaper_app/constants/color_constants.dart';
+import 'package:magic_epaper_app/util/orientation_util.dart';
 import 'package:magic_epaper_app/view/widget/common_scaffold_widget.dart';
 
 //TODO add Language Support and Dark mode support here
@@ -18,15 +19,8 @@ class SettingsScreenState extends State<SettingsScreen> {
 
   @override
   void initState() {
-    _setOrientation();
+    setPortraitOrientation();
     super.initState();
-  }
-
-  void _setOrientation() {
-    SystemChrome.setPreferredOrientations([
-      DeviceOrientation.portraitUp,
-      DeviceOrientation.portraitDown,
-    ]);
   }
 
   @override
