+++ lib/view/widget/navigation_drawer.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:magic_epaper_app/constants/color_constants.dart';
 import 'package:magic_epaper_app/constants/string_constants.dart';
 import 'package:share_plus/share_plus.dart';
-import 'package:magic_epaper_app/constants/url_constant.dart';
+import 'package:magic_epaper_app/util/url_util.dart';
 
 class AppDrawer extends StatefulWidget {
   final int selectedIndex;
@@ -165,7 +165,7 @@ class _AppDrawerState extends State<AppDrawer> {
         if (externalLink != null) {
           openUrl(context, externalLink);
         } else if (shareText != null) {
-          Share.share(shareText);
+          SharePlus.instance.share(ShareParams(text: shareText));
         } else {
           Navigator.pushNamedAndRemoveUntil(
             context,
